copy u-boot from SDcard to DDR
Note: reference 《Android_Exynos4412_iROM_Secure_Booting_Guide_Ver.1.00.00.pdf》P21
Warning: The frequency of clocks supplied to SDMMC and eMMC are 20Mhz at the Booting time. MPLL is the source of these clocks.
Warning: If SDMMC or eMMC is chosen as the booting device, the copy functions for SDMMC or eMMC would be available in BL1 and BL2.
         If you use the copy function, please do not change the clocks for SDMMC or eMMC. Additionally do not change the configuration
         of PLL related to SDMMC or eMMC. Proper booting operations could not be guaranteed under illegal clock changes.

path:
	modified:   arch/arm/mach-exynos/clock_init_exynos4412.c
	modified:   arch/arm/mach-exynos/lowlevel_init.c
	modified:   include/configs/tiny4412.h

Signed-off-by: ap0904225 <ap09042250@qq.com>